Combine the SpiralWiper into the Wiper.
