+++ src/shapes/Wiper.js
@@ -5,12 +5,18 @@ import Shape, { shapeOptions } from './Shape'
 const options = {
   ...shapeOptions,
   ...{
-    wiperAngleDeg: {
-      title: 'Wiper angle',
+    wiperType: {
+      title: 'Type',
+      type: 'dropdown',
+      choices: ['Lines', 'Spiral'],
     },
     wiperSize: {
       title: 'Wiper size',
       min: 1
+    },
+    wiperAngleDeg: {
+      title: 'Wiper angle',
+      isVisible: (state) => { return state.wiperType === 'Lines' },
     },
   }
 }
@@ -74,6 +80,165 @@ function nearEnough(end, point) {
   return false
 }
 
+function spiralVertices(state) {
+  // Determine the max radius
+  let maxRadius = state.machine.maxRadius
+  if (state.machine.rectangular) {
+    const halfHeight = (state.machine.maxY - state.machine.minY)/2.0
+    const halfWidth = (state.machine.maxX - state.machine.minX)/2.0
+    maxRadius = Math.sqrt(Math.pow(halfHeight, 2.0) + Math.pow(halfWidth, 2.0))
+  }
+
+  var vertices = []
+
+  let emergencyBreak = 0
+  let radius = 0.1
+  let angle = 0
+  const arcLength = 1.0
+  const radiusPerAngle = state.shape.wiperSize / (2.0 * Math.PI)
+  while (radius <= maxRadius) {
+    // Save where we are right now.
+    vertices.push(new Victor(radius * Math.cos(angle), radius * Math.sin(angle)))
+
+    // We want to have the next point be about the right arc length.
+    let deltaAngle = arcLength / radius * 2.0 * Math.PI
+
+    // Limit this at small radii
+    deltaAngle = Math.min(deltaAngle, 0.1)
+
+    // Update for the next point.
+    angle += deltaAngle
+    radius += deltaAngle * radiusPerAngle
+
+    emergencyBreak += 1
+    if (emergencyBreak > 1e9) {
+      // If we get to this point: bail. We did something terrible.
+      break
+    }
+  }
+
+  return vertices
+}
+
+function linearVertices(state) {
+  var outputVertices = []
+
+  // Get the angle between 0,180
+  let angle = (180.0 - (state.shape.wiperAngleDeg % 360)) % 180.0
+  if (angle < 0.0) {
+    angle += 180.0
+  }
+  angle = degToRad(angle)
+
+  // Start with the defaults for 0,45
+  let height = 1
+  let width = 1
+  let machine = state.machine
+  if (machine.rectangular) {
+    height = machine.maxY - machine.minY
+    width = machine.maxX - machine.minX
+  } else {
+    height = machine.maxRadius * 2.0
+    width = height
+  }
+
+  let startLocation = Victor(-width/2.0, height/2.0)
+  let cosa = Math.cos(angle)
+  let sina = Math.sin(angle)
+
+  // These can be zero, but infinity isn't great for out math, so let's just clip it.
+  if (Math.abs(cosa) < 1.0e-10) {
+    cosa = 1.0e-10
+  }
+  if (Math.abs(sina) < 1.0e-10) {
+    sina = 1.0e-10
+  }
+  let orig_delta_w = Victor(state.shape.wiperSize / cosa, 0.0)
+  let orig_delta_h = Victor(0.0, -state.shape.wiperSize / sina)
+
+  if (angle > Math.PI/4.0 && angle < 0.75 * Math.PI) {
+    // flip the logic of x,y
+    let temp = orig_delta_w.clone()
+    orig_delta_w = orig_delta_h.clone()
+    orig_delta_h = temp
+  }
+  if (angle > Math.PI/2.0) {
+    startLocation = Victor(-width/2.0, -height/2.0)
+    orig_delta_w = orig_delta_w.clone().multiply(Victor(-1.0, -1.0))
+    orig_delta_h = orig_delta_h.clone().multiply(Victor(-1.0, -1.0))
+  }
+  let delta_w = orig_delta_w
+  let delta_h = orig_delta_h
+  let endLocation = startLocation.clone().multiply(Victor(-1.0, -1.0))
+  outputVertices.push(startLocation)
+  let nextWidthPoint = startLocation
+  let nextHeightPoint = startLocation
+
+  let emergency_break = 0
+  while (emergency_break < 1000) {
+    emergency_break += 1
+
+    // "right"
+    nextWidthPoint = nextWidthPoint.clone().add(delta_w)
+    if (outOfBounds(nextWidthPoint, width, height)) {
+      let corner = boundPoint(nextWidthPoint.clone().subtract(delta_w), nextWidthPoint, width/2.0, height/2.0)
+      outputVertices.push(corner)
+      if (nearEnough(endLocation, corner)) {
+        break
+      }
+      nextWidthPoint = boundPoint(nextHeightPoint, nextWidthPoint, width/2.0, height/2.0)
+      delta_w = orig_delta_h
+    }
+    outputVertices.push(nextWidthPoint)
+    if (nearEnough(endLocation, nextWidthPoint)) {
+      break
+    }
+
+    // "down-left"
+    nextHeightPoint = nextHeightPoint.clone().add(delta_h)
+    if (outOfBounds(nextHeightPoint, width, height)) {
+      nextHeightPoint = boundPoint(nextWidthPoint, nextHeightPoint, width/2.0, height/2.0)
+      delta_h = orig_delta_w
+    }
+    outputVertices.push(nextHeightPoint)
+    if (nearEnough(endLocation, nextHeightPoint)) {
+      break
+    }
+
+    // "down"
+    nextHeightPoint = nextHeightPoint.clone().add(delta_h)
+    outputVertices.push(nextHeightPoint)
+    if (nearEnough(endLocation, nextHeightPoint)) {
+      break
+    }
+    if (outOfBounds(nextHeightPoint, width, height)) {
+      let corner = boundPoint(nextHeightPoint.clone().subtract(delta_h), nextHeightPoint, width/2.0, height/2.0)
+      outputVertices.push(corner)
+      if (nearEnough(endLocation, corner)) {
+        break
+      }
+      nextHeightPoint = boundPoint(nextWidthPoint, nextHeightPoint, width/2.0, height/2.0)
+      delta_h = orig_delta_w
+    }
+    outputVertices.push(nextHeightPoint)
+    if (nearEnough(endLocation, nextHeightPoint)) {
+      break
+    }
+
+    // "up-right"
+    nextWidthPoint = nextWidthPoint.clone().add(delta_w)
+    outputVertices.push(nextWidthPoint)
+    if (nearEnough(endLocation, nextWidthPoint)) {
+      break
+    }
+    if (outOfBounds(nextWidthPoint, width, height)) {
+      nextWidthPoint = boundPoint(nextHeightPoint, nextWidthPoint, width/2.0, height/2.0)
+      delta_w = orig_delta_h
+    }
+  }
+  return outputVertices
+}
+
 export default class Wiper extends Shape {
   constructor() {
     super('Wiper')
@@ -86,6 +251,7 @@ export default class Wiper extends Shape {
         type: 'wiper',
         wiperAngleDeg: 15,
         wiperSize: 12,
+        wiperType: 'Lines',
         startingSize: 1,
         canChangeSize: false,
         repeatEnabled: false,
@@ -96,122 +262,11 @@ export default class Wiper extends Shape {
   }
 
   getVertices(state) {
-    var outputVertices = []
-
-    // Get the angle between 0,180
-    let angle = (180.0 - (state.shape.wiperAngleDeg % 360)) % 180.0
-    if (angle < 0.0) {
-      angle += 180.0
-    }
-    angle = degToRad(angle)
-
-    // Start with the defaults for 0,45
-    let height = 1
-    let width = 1
-    let machine = state.machine
-    if (machine.rectangular) {
-      height = machine.maxY - machine.minY
-      width = machine.maxX - machine.minX
+    if (state.shape.wiperType === 'Lines') {
+      return linearVertices(state)
     } else {
-      height = machine.maxRadius * 2.0
-      width = height
+      return spiralVertices(state)
     }
-
-    let startLocation = Victor(-width/2.0, height/2.0)
-    let cosa = Math.cos(angle)
-    let sina = Math.sin(angle)
-
-    // These can be zero, but infinity isn't great for out math, so let's just clip it.
-    if (Math.abs(cosa) < 1.0e-10) {
-      cosa = 1.0e-10
-    }
-    if (Math.abs(sina) < 1.0e-10) {
-      sina = 1.0e-10
-    }
-    let orig_delta_w = Victor(state.shape.wiperSize / cosa, 0.0)
-    let orig_delta_h = Victor(0.0, -state.shape.wiperSize / sina)
-
-    if (angle > Math.PI/4.0 && angle < 0.75 * Math.PI) {
-      // flip the logic of x,y
-      let temp = orig_delta_w.clone()
-      orig_delta_w = orig_delta_h.clone()
-      orig_delta_h = temp
-    }
-    if (angle > Math.PI/2.0) {
-      startLocation = Victor(-width/2.0, -height/2.0)
-      orig_delta_w = orig_delta_w.clone().multiply(Victor(-1.0, -1.0))
-      orig_delta_h = orig_delta_h.clone().multiply(Victor(-1.0, -1.0))
-    }
-    let delta_w = orig_delta_w
-    let delta_h = orig_delta_h
-    let endLocation = startLocation.clone().multiply(Victor(-1.0, -1.0))
-    outputVertices.push(startLocation)
-    let nextWidthPoint = startLocation
-    let nextHeightPoint = startLocation
-
-    let emergency_break = 0
-    while (emergency_break < 1000) {
-      emergency_break += 1
-
-      // "right"
-      nextWidthPoint = nextWidthPoint.clone().add(delta_w)
-      if (outOfBounds(nextWidthPoint, width, height)) {
-        let corner = boundPoint(nextWidthPoint.clone().subtract(delta_w), nextWidthPoint, width/2.0, height/2.0)
-        outputVertices.push(corner)
-        if (nearEnough(endLocation, corner)) {
-          break
-        }
-        nextWidthPoint = boundPoint(nextHeightPoint, nextWidthPoint, width/2.0, height/2.0)
-        delta_w = orig_delta_h
-      }
-      outputVertices.push(nextWidthPoint)
-      if (nearEnough(endLocation, nextWidthPoint)) {
-        break
-      }
-
-      // "down-left"
-      nextHeightPoint = nextHeightPoint.clone().add(delta_h)
-      if (outOfBounds(nextHeightPoint, width, height)) {
-        nextHeightPoint = boundPoint(nextWidthPoint, nextHeightPoint, width/2.0, height/2.0)
-        delta_h = orig_delta_w
-      }
-      outputVertices.push(nextHeightPoint)
-      if (nearEnough(endLocation, nextHeightPoint)) {
-        break
-      }
-
-      // "down"
-      nextHeightPoint = nextHeightPoint.clone().add(delta_h)
-      outputVertices.push(nextHeightPoint)
-      if (nearEnough(endLocation, nextHeightPoint)) {
-        break
-      }
-      if (outOfBounds(nextHeightPoint, width, height)) {
-        let corner = boundPoint(nextHeightPoint.clone().subtract(delta_h), nextHeightPoint, width/2.0, height/2.0)
-        outputVertices.push(corner)
-        if (nearEnough(endLocation, corner)) {
-          break
-        }
-        nextHeightPoint = boundPoint(nextWidthPoint, nextHeightPoint, width/2.0, height/2.0)
-        delta_h = orig_delta_w
-      }
-      outputVertices.push(nextHeightPoint)
-      if (nearEnough(endLocation, nextHeightPoint)) {
-        break
-      }
-
-      // "up-right"
-      nextWidthPoint = nextWidthPoint.clone().add(delta_w)
-      outputVertices.push(nextWidthPoint)
-      if (nearEnough(endLocation, nextWidthPoint)) {
-        break
-      }
-      if (outOfBounds(nextWidthPoint, width, height)) {
-        nextWidthPoint = boundPoint(nextHeightPoint, nextWidthPoint, width/2.0, height/2.0)
-        delta_w = orig_delta_h
-      }
-    }
-    return outputVertices
   }
 
   getOptions() {
